Button event calculations improvements

+++ src/texture.h
@@ -9,12 +9,12 @@ typedef struct JUN_TextureData JUN_TextureData;
 struct JUN_TextureData
 {
     uint8_t id;
-    uint32_t x;
-    uint32_t y;
-    uint32_t width;
-    uint32_t height;
-    uint32_t image_width;
-    uint32_t image_height;
+    float x;
+    float y;
+    float width;
+    float height;
+    float image_width;
+    float image_height;
 };
 
 JUN_Texture *JUN_TextureCreateContext(uint32_t view_width, uint32_t view_height, uint8_t offset);
